refactor(ios): remove dead poster components
MobileMediaPosterButton, MobilePosterCard, MobileMediaRow and
MobileMediaGridView had no live callers (the grids use
MobileRecentlyAddedCard). Delete them; preserve the still-used
ItemContextMenu by moving it to its own file.
